pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# location of package relative to the repository root (e.g. alternatively "src/my_library")
# note: the package name 'my_library' can be different from the project name on PyPI
#       the project name (PyPI) is used for pip install, whereas the package name is used for imports
packages = ["autoeval"] 

[tool.hatch.build.targets.sdist]
exclude = ["notebooks", "basic"]

[project]
name = "autoeval"
description = "Package to automate some postprocessing functionalities of MIKE simulations."
version = "0.0.1"
authors = [
  { name = "Faro Schäfer", email = "fsch@dhigroup.com" },
]
requires-python = ">= 3.10"
readme = "README.md"
classifiers = [
  "Programming Language :: Python"
]

dependencies = [
    "numpy",
    "mikeio",
    "pandas",
    "modelskill", 
    "python-pptx",
    "matplotlib",
    "contextily",
]